Decouple Idna conversion
