+++ src/TopLevelDomainsTest.php
@@ -193,7 +193,7 @@ final class TopLevelDomainsTest extends TestCase
     {
         $resolved = $this->topLevelDomains->resolve('foo.de');
 
-        self::assertTrue($resolved->domain()->isIDNA2008());
+        self::assertTrue($resolved->domain()->isIdna2008());
 
         $collection = TopLevelDomains::fromPath(
             dirname(__DIR__).'/test_data/root_zones.dat',
@@ -203,7 +203,7 @@ final class TopLevelDomainsTest extends TestCase
         $domain = Domain::fromIDNA2003('foo.de');
         $resolved = $collection->resolve($domain);
 
-        self::assertFalse($resolved->domain()->isIDNA2008());
+        self::assertFalse($resolved->domain()->isIdna2008());
     }
 
     /**
